Colocate commit chart and move roundMax
