+++ src/utils/numbers/numbers.test.ts
@@ -1,4 +1,4 @@
-import { roundDecimal, cryptoRandom } from ".";
+import { roundDecimal, cryptoRandom, roundMax } from ".";
 
 describe("roundDecimal", () => {
   it("correctly rounds a decimal to the specified number of places", () => {
@@ -17,5 +17,15 @@ describe("cryptoRandom", () => {
     const randomNumber = cryptoRandom();
     expect(randomNumber).toBeGreaterThanOrEqual(0);
     expect(randomNumber).toBeLessThan(1);
+  });
+});
+
+describe("roundMax", () => {
+  it("properly rounds numbers", () => {
+    expect(roundMax(8)).toBe(10); // 0 <= x < 100
+    expect(roundMax(147)).toBe(150); // 100 <= x < 500
+    expect(roundMax(712)).toBe(800); // 500 <= x < 1000
+    expect(roundMax(1320)).toBe(1500); // 1000 <= x < 5000
+    expect(roundMax(6430)).toBe(7000); // 5000 <= x
   });
 });
